Improve encapsulation of ecommerce check warnings
* Warning messages for unsupported e-commerce states are now
encapsulated in the state checking functions and only issued if the warn
parameter is passed in with a value of true. This removes the repeated
conditionals and strings from other locations in the code base.
* Added additional warning message to explain functionality not working
when enhanced ecommerce is enabled and ecommerce only commands are no
longer supported.
* Added two new test cases for ensuring the various commands issues
correct warnings based on state.