Add support for multiple trackers everywhere
All analytics.js calls now support multiple trackers with backwards
compatibility of the existing behavior. In order to enable a tracker
when multiple accounts are set the 'trackEcommerce' property needs to be
defined as 'true' on the tracker object. If none of the accounts are set
then the multiple tracker function falls back to single tracker behavior
as is the previous behavior. This change will not break existing
implementations.